5th Day of the Project
Today I finished the See fragment. For now on you can write and read the data from the Firebase. My next goal is to add a delete function, so the user will able to delete data from the fragment if click the recyclerview there will be appear a delete button and user can delete any data from the list.